requirements.txt
streamlit>=1.32.0
requests>=2.31.0
dnspython>=2.4.2
python-whois>=0.8.0
beautifulsoup4>=4.12.2
lxml>=5.1.0
urllib3>=2.0.0
python-nmap>=0.7.1
reportlab>=4.1.0
plotly>=5.18.0
pandas>=2.2.0
colorama>=0.4.6
tqdm>=4.66.0
aiohttp>=3.9.0
asyncio>=3.4.3
jinja2>=3.1.3
markdown>=3.5.2
